Password protect for deleting

+++ scouting_app/lib/settings.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_form_builder/flutter_form_builder.dart';
 import 'package:scoutingapp/FileUtils.dart';
 import 'package:scoutingapp/scouting_icons_icons.dart';
 
@@ -17,6 +18,9 @@ class MySettingPage extends StatefulWidget {
 }
 
 class MySettingPageState extends State<MySettingPage> {
+  final GlobalKey<FormBuilderState> _fbKey = GlobalKey<FormBuilderState>();
+  final GlobalKey<FormBuilderState> _formKey = GlobalKey<FormBuilderState>();
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -103,22 +107,58 @@ class MySettingPageState extends State<MySettingPage> {
                     context: context,
                     builder: (BuildContext context) {
                       return AlertDialog(
+                        shape: RoundedRectangleBorder(
+                          borderRadius: BorderRadius.all(
+                            Radius.circular(10),
+                          ),
+                        ),
                         title: const Text("Confirm"),
-                        content: const Text(
-                            "Are you sure you wish to clear all forms and schedule?"),
+                        content: Container(
+                          height: 129,
+                          child: Column(
+                            children: <Widget>[
+                              Text(
+                                  "Are you sure you wish to clear all forms and schedule?"),
+                              SizedBox(
+                                height: 10,
+                              ),
+                              FormBuilder(
+                                key: _formKey,
+                                child: FormBuilderTextField(
+                                  attribute: "password",
+                                  validators: [
+                                    FormBuilderValidators.required(),
+                                    // ignore: missing_return
+                                    (val) {
+                                      if (val != "CRyptonite2001")
+                                        return "Incorrect Password";
+                                    },
+                                  ],
+                                  maxLines: 1,
+                                  obscureText: true,
+                                  decoration:
+                                      InputDecoration(labelText: "Password"),
+                                ),
+                              )
+                            ],
+                          ),
+                        ),
                         actions: <Widget>[
                           FlatButton(
                             child: Text("CONFIRM"),
                             onPressed: () {
-                              FileUtils.deleteSchedule();
-                              FileUtils.deletePitForm();
-                              FileUtils.deleteForm();
-                              Navigator.pushAndRemoveUntil(
-                                context,
-                                MaterialPageRoute(
-                                    builder: (context) => MyHomePage()),
-                                (Route<dynamic> route) => false,
-                              );
+                              print(_formKey);
+                              if (_formKey.currentState.saveAndValidate()) {
+                                try {
+                                  FileUtils.deleteData();
+                                } catch (e) {}
+                                Navigator.pushAndRemoveUntil(
+                                  context,
+                                  MaterialPageRoute(
+                                      builder: (context) => MyHomePage()),
+                                  (Route<dynamic> route) => false,
+                                );
+                              }
                             },
                           ),
                           FlatButton(
@@ -145,19 +185,56 @@ class MySettingPageState extends State<MySettingPage> {
                     builder: (BuildContext context) {
                       return AlertDialog(
                         title: const Text("Confirm"),
-                        content: const Text(
-                            "Are you sure you wish to clear ALL data?"),
+                        shape: RoundedRectangleBorder(
+                          borderRadius: BorderRadius.all(
+                            Radius.circular(10),
+                          ),
+                        ),
+                        content: Container(
+                          height: 129,
+                          child: Column(
+                            children: <Widget>[
+                              Text("Are you sure you wish to clear ALL data?"),
+                              SizedBox(
+                                height: 10,
+                              ),
+                              FormBuilder(
+                                key: _fbKey,
+                                child: FormBuilderTextField(
+                                  attribute: "password",
+                                  validators: [
+                                    FormBuilderValidators.required(),
+                                    // ignore: missing_return
+                                    (val) {
+                                      if (val != "CRyptonite2001")
+                                        return "Incorrect Password";
+                                    },
+                                  ],
+                                  maxLines: 1,
+                                  obscureText: true,
+                                  decoration:
+                                      InputDecoration(labelText: "Password"),
+                                ),
+                              )
+                            ],
+                          ),
+                        ),
                         actions: <Widget>[
                           FlatButton(
                             child: Text("CONFIRM"),
                             onPressed: () {
-                              FileUtils.deleteData();
-                              Navigator.pushAndRemoveUntil(
-                                context,
-                                MaterialPageRoute(
-                                    builder: (context) => MyHomePage()),
-                                (Route<dynamic> route) => false,
-                              );
+                              print(_fbKey);
+                              if (_fbKey.currentState.saveAndValidate()) {
+                                try {
+                                  FileUtils.deleteData();
+                                } catch (e) {}
+                                Navigator.pushAndRemoveUntil(
+                                  context,
+                                  MaterialPageRoute(
+                                      builder: (context) => MyHomePage()),
+                                  (Route<dynamic> route) => false,
+                                );
+                              }
                             },
                           ),
                           FlatButton(
